watchdog: use the kernel api to stop the hw watchdog
If not stopping the watchdog via API it will continue running and not
be triggered by procd any more. After timing out it reset the board.
A stop requires to close the wdt file and for many drivers e.g. ar71xx
it also requires to write 'V' before closing the file descriptor.

Signed-off-by: Alexander Couzens <lynxis@fe80.eu>